# fix(l1, l2): use `BTreeMap` for `InMemoryTrieDB` and optimize stateless execution by avoiding state cloning