--- a/.github/workflows/main_prover_l1.yaml
+++ b/.github/workflows/main_prover_l1.yaml
@@ -64,9 +64,9 @@ jobs:
           # RPC_URL: $${{ secrets.ETHREX_L2_RPC_URL }}
           RPC_URL: "http://eth-testnet-rpc:8547"
         run: |
-          cp fixtures/cache/rpc_prover/cache_hoodi_1130549.bin cmd/ethrex_replay
+          cp fixtures/cache/rpc_prover/cache_hoodi_1132599.bin cmd/ethrex_replay
           cd cmd/ethrex_replay
-          make prove-${{ matrix.backend }}-gpu-ci BLOCK_NUMBER=1130549 NETWORK=hoodi
+          make prove-${{ matrix.backend }}-gpu-ci BLOCK_NUMBER=1132599 NETWORK=hoodi
 
       - name: Store benchmark result
         uses: benchmark-action/github-action-benchmark@v1