--- a/.github/workflows/pr-main_l1.yaml
+++ b/.github/workflows/pr-main_l1.yaml
@@ -251,9 +251,9 @@ jobs:
           # RPC_URL: $${{ secrets.ETHREX_L2_RPC_URL }}
           RPC_URL: "http://45.76.185.153:8545"
         run: |
-          cp fixtures/cache/rpc_prover/cache_hoodi_1130549.bin cmd/ethrex_replay
+          cp fixtures/cache/rpc_prover/cache_hoodi_1132599.bin cmd/ethrex_replay
           cd cmd/ethrex_replay
-          make execute BLOCK_NUMBER=1130549 NETWORK=hoodi
+          make execute BLOCK_NUMBER=1132599 NETWORK=hoodi
 
   # The purpose of this job is to add it as a required check in GitHub so that we don't have to add every individual job as a required check
   all-tests: