--- a/Cargo.toml
+++ b/Cargo.toml
@@ -115,7 +115,7 @@ spawned-rt = "0.3.0"
 lambdaworks-crypto = "0.11.0"
 tui-logger = { version = "0.17.3", features = ["tracing-support"] }
 rayon = "1.10.0"
-rkyv = "0.8.10"
+rkyv = { version = "0.8.10", features = ["std"] }
 
 [patch.crates-io]
 secp256k1 = { git = "https://github.com/sp1-patches/rust-secp256k1", tag = "patch-0.29.1-sp1-5.0.0" }