--- a/crates/common/rkyv_utils.rs
+++ b/crates/common/rkyv_utils.rs
@@ -63,7 +63,9 @@ impl Hash for ArchivedH160Wrapper {
     }
 }
 
-#[derive(Archive, Serialize, Deserialize, Clone, Copy, Debug, PartialEq, Eq, Hash)]
+#[derive(
+    Archive, Serialize, Deserialize, Clone, Copy, Debug, PartialEq, Eq, Hash, PartialOrd, Ord,
+)]
 #[rkyv(remote = H256)]
 pub struct H256Wrapper([u8; 32]);
 
@@ -76,6 +78,18 @@ impl From<H256Wrapper> for H256 {
 impl PartialEq for ArchivedH256Wrapper {
     fn eq(&self, other: &Self) -> bool {
         self.0 == other.0
+    }
+}
+
+impl PartialOrd for ArchivedH256Wrapper {
+    fn partial_cmp(&self, other: &Self) -> Option<std::cmp::Ordering> {
+        Some(self.cmp(other))
+    }
+}
+
+impl Ord for ArchivedH256Wrapper {
+    fn cmp(&self, other: &Self) -> std::cmp::Ordering {
+        self.0.cmp(&other.0)
     }
 }