**Motivation**

Stateless execution with SP1 is stable, but it still takes too much time to execute and prove compared with other tools.

1. We were aware that we were cloning the entire state on each trie rebuild (state and account storage ones).
2. `HashMap` insertions are expensive for the prover. By changing the data structure used under the hood, we can avoid these operations.

**Description**

- Use a `BTreeMap` as the `InMemoryTrieDB` inner storage instead of a `HashMap`.
- Pass a reference to the state nodes map instead of cloning it every time we want to build a trie.

**Results**

*`mainnet`'s 23276522  block Execution Time with SP1 Before (145s)*

```
2025-09-02T17:16:17.815246Z  INFO ethrex_replay::fetcher: Retrieving execution data for block 23276522 (329 block behind latest)
2025-09-02T17:16:17.821839Z  INFO ethrex_replay::fetcher: Getting block 23276522 data from cache
2025-09-02T17:16:17.822414Z  INFO ethrex_replay::bench: Starting prover program
2025-09-02T17:16:17.824406Z  WARN sp1_sdk::env: SP1_PROVER environment variable not set, defaulting to 'cpu'
stderr: WARNING: Using insecure random number generator.
2025-09-02T17:18:45.292679Z  INFO execute: sp1_prover: gas: 7777466726
2025-09-02T17:18:45.298342Z  INFO ethrex_prover_lib::backends::sp1: Successfully executed SP1 program in 145.31s
2025-09-02T17:18:45.299427Z  INFO ethrex_replay::bench: Total gas from block/s: 22330686
```

*`mainnet`'s 23276522  block Execution Time with SP1 After (35s)*

```
2025-09-02T18:56:28.782876Z  INFO ethrex_replay::fetcher: Retrieving execution data for block 23276522 (827 block behind latest)
2025-09-02T18:56:28.789067Z  INFO ethrex_replay::fetcher: Getting block 23276522 data from cache
2025-09-02T18:56:28.789627Z  INFO ethrex_replay::bench: Starting prover program
2025-09-02T18:56:28.791437Z  WARN sp1_sdk::env: SP1_PROVER environment variable not set, defaulting to 'cpu'
stderr: WARNING: Using insecure random number generator.
2025-09-02T18:57:05.758187Z  INFO execute: sp1_prover: gas: 1629109439
2025-09-02T18:57:05.760484Z  INFO ethrex_prover_lib::backends::sp1: Successfully executed SP1 program in 34.81s
2025-09-02T18:57:05.761603Z  INFO ethrex_replay::bench: Total gas from block/s: 22330686
```

*`mainnet`'s 23276522  block Execution with SP1 Flamegraph Before (`get_storage_slot` = ~80%)*

<img width="1512" height="457" alt="image" src="https://github.com/user-attachments/assets/b0bd9b36-a998-4228-ac9e-8b583fd986d5" />

*`mainnet`'s 23276522  block Execution with SP1 Flamegraph After (`get_storage_slot` = ~18%)*

<img width="1512" height="437" alt="image" src="https://github.com/user-attachments/assets/4e90f631-3e14-47ac-a3dd-ba4871237b83" />